+++ src/app/expenses/add/page.tsx
@@ -1,59 +1,14 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
 import ExpenseForm from "../components/ExpenseForm";
 import MessageDisplay from "../../components/MessageDisplay";
-import { addExpense } from "../../api";
+import useExpenseForm from "../../hooks/useExpenseForm";
 import Header from "@/app/components/Header";
-
-interface Expense {
-  id?: number;
-  title: string;
-  description: string;
-  amount: number;
-  date: Date;
-  username?: string;
-}
 
 const AddExpensePage: React.FC = () => {
   const router = useRouter();
-  const [message, setMessage] = useState<string | null>(null);
-  const [isSuccess, setIsSuccess] = useState<boolean>(false);
-
-  const username = sessionStorage.getItem("username");
-
-  useEffect(() => {
-    if (!username) {
-      router.push("/");
-    }
-  }, [username]);
-
-  const handleAdd = async (expense: Expense) => {
-    if (!username) {
-      setMessage("User not found. Please log in again.");
-      setIsSuccess(false);
-      return;
-    }
-
-    const expenseData = {
-      ...expense,
-      username,
-    };
-
-    try {
-      await addExpense(expenseData);
-      setMessage("Expense added successfully!");
-      setIsSuccess(true);
-      setTimeout(() => router.push("/expenses"), 1000);
-    } catch (error) {
-      setMessage("Failed to add expense. Please try again.");
-      setIsSuccess(false);
-      setTimeout(() => {
-        setMessage(null);
-      }, 1000);
-    }
-  };
+  const { message, isSuccess, handleSubmit } = useExpenseForm();
 
   return (
     <div>
@@ -67,7 +22,6 @@ const AddExpensePage: React.FC = () => {
         <div className="bg-white p-8 rounded-lg shadow-md text-left max-w-xl w-full">
           <div className="flex justify-between items-center mb-4">
             <h1 className="text-3xl font-bold text-gray-800">Add Expense</h1>
-            {/* Back button */}
             <button
               onClick={() => router.back()}
               className="bg-gray-600 text-white px-4 py-2 rounded hover:bg-gray-800 transition-colors"
@@ -75,7 +29,7 @@ const AddExpensePage: React.FC = () => {
               <strong>{">"}</strong>
             </button>
           </div>
-          <ExpenseForm onSubmit={handleAdd} isEdit={false} />
+          <ExpenseForm onSubmit={handleSubmit} isEdit={false} />
         </div>
       </div>
     </div>
